move demo.py into demo/test_logging.py
Done to make docstring with explanation more visible and to avoid flake8
errors the docstring caused.
